Lots of small fixes:  - fixed meads_milkweed growth stages  - tall plant worldgen takes maxHeight into account  - tweaked cactus worldgen to be less dense  - plants now use Forge DecorateBiomeEvent  - tall grass now uses ColorizerGrass instead of ColorizerFoliage  - shrank licorice fern model  - fixed translations for item models

+++ src/main/java/net/dries007/tfc/util/Helpers.java
@@ -95,7 +95,7 @@ public final class Helpers
                     if (world.isAirBlock(pos.up()) &&
                         plant.isValidLocation(temp, ChunkDataTFC.getRainfall(world, pos.up()), world.getLightFromNeighbors(pos.up())) &&
                         plant.isValidGrowthTemp(temp) &&
-                        rand.nextFloat() < plantBlock.getGrowthRate(world, pos.up()))
+                        rand.nextDouble() < plantBlock.getGrowthRate(world, pos.up()))
                     {
                         world.setBlockState(pos.up(), plantBlock.getDefaultState());
                     }
